requirements.txt
fastapi==0.115.8
uvicorn==0.35.0
pydantic[email]>=2.0.0
SQLAlchemy==2.0.42
PyMySQL==1.1.2
python-dotenv==1.1.1
python-jose[cryptography]
passlib[bcrypt]
fastapi-mail>=1.4.1
alembic
python-multipart==0.0.20
